+++ lib/autumn/theme.ex
@@ -23,7 +23,10 @@ defmodule Autumn.Theme do
              underline: false,
              bold: false,
              italic: false,
-             strikethrough: false
+             strikethrough: false,
+             display: nil,
+             width: nil,
+             transition: nil
            },
            "variable.builtin" => %Autumn.Theme.Style{
              fg: "#0550ae",
@@ -31,7 +34,10 @@ defmodule Autumn.Theme do
              underline: false,
              bold: false,
              italic: false,
-             strikethrough: false
+             strikethrough: false,
+             display: nil,
+             width: nil,
+             transition: nil
            },
            "character" => %Autumn.Theme.Style{
              fg: "#0a3069",
@@ -39,7 +45,10 @@ defmodule Autumn.Theme do
              underline: false,
              bold: false,
              italic: false,
-             strikethrough: false
+             strikethrough: false,
+             display: nil,
+             width: nil,
+             transition: nil
            },
            ...
          }
@@ -103,8 +112,19 @@ defmodule Autumn.Theme.Style do
           underline: boolean(),
           bold: boolean(),
           italic: boolean(),
-          strikethrough: boolean()
+          strikethrough: boolean(),
+          display: nil | String.t(),
+          width: nil | String.t(),
+          transition: nil | String.t()
         }
 
-  defstruct fg: nil, bg: nil, underline: false, bold: false, italic: false, strikethrough: false
+  defstruct fg: nil,
+            bg: nil,
+            underline: false,
+            bold: false,
+            italic: false,
+            strikethrough: false,
+            display: nil,
+            width: nil,
+            transition: nil
 end